pyproject.toml
[project]
name = "experiment_langchain_agent"
version = "0.0.1"
requires-python = ">=3.11"
dependencies = [
    "langchain == 0.0.178",
    "gpt4all == 0.2.3",
    "wikipedia == 1.4.0",
    "google-generativeai",
    "openai",
    "duckduckgo-search == 2.8.6",
    "sentence-transformers",
    "chromadb",
    "pdfminer.six",
    "python-dotenv",
]


# https://setuptools.pypa.io/en/latest/userguide/quickstart.html#development-mode
# pip install --editable .